+++ R/est_ggls.R
@@ -116,6 +116,8 @@ pggls <- function(formula, data, subset, na.action,
     coef.names <- names(coef(plm.model))
     K <- length(coef.names)
 
+    
+    ## TODO: for FD models respecting time-wise diffing: need dedicated procedure to construct index
     if (model.name == "fd") {
     ## eliminate first year in indices
         nt <- pdim$Tint$nt[-1L]
@@ -160,6 +162,7 @@ pggls <- function(formula, data, subset, na.action,
         other.names <- time.names
         groupsdim <- Ti
     }
+
     myord <- order(cond, other)
     X <- model.matrix(plm.model)[myord, , drop = FALSE]
     commonpars <- intersect(coef.names, colnames(X))
@@ -202,7 +205,7 @@ pggls <- function(formula, data, subset, na.action,
     }
 
     if (balanced) {
-        tres <- split(resid, cond)
+        tres <- collapse::gsplit(resid, cond)
         tres <- lapply(tres, function(x) outer(x, x))
         tres <- array(unlist(tres), dim = c(nother, nother, ncond)) # make array so rowMeans can be used
       
@@ -214,8 +217,9 @@ pggls <- function(formula, data, subset, na.action,
                       dimnames = list(other.names, other.names, cond.names))
         
         # split data by cond
-        resid.list <- split(resid, cond)
-        other.list <- split(other, cond)
+        cond.GRP <- collapse::GRP(cond)
+        resid.list <- collapse::gsplit(resid, cond.GRP)
+        other.list <- collapse::gsplit(other, cond.GRP)
         
         for (i in seq_len(ncond)) {
             ut <- resid.list[[i]]