+++ R/tool_transformations.R
@@ -132,7 +132,7 @@ as.matrix.pseries <- function(x, idbyrow = TRUE, ...){
     id <- index[[1L]]
     time <- index[[2L]]
     time.names <- levels(time)
-    x <- split(data.frame(x, time), id)
+    x <- collapse::rsplit(data.frame(x, time), id)
     x <- lapply(x, function(x){
         rownames(x) <- x[ , 2L]
         x[ , -2L, drop = FALSE]
@@ -765,24 +765,72 @@ lead <- function(x, k = 1L, ...) {
 #' @export lag
 lag.pseries <- function(x, k = 1L, shift = c("time", "row"), ...) {
   shift <- match.arg(shift)
-  res <- if(shift == "time") lagt.pseries(x = x, k = k, ...) else lagr.pseries(x = x, k = k, ...)
-  return(res)
+  
+  if (shift == "time") {
+    if (!isTRUE(getOption("plm.fast"))) {
+      res <- lagt.pseries(x = x, k = k, ...) # base R
+    } else {
+      if (!isTRUE(getOption("plm.fast.pkg.collapse"))) stop(txt.no.collapse, call. = FALSE)
+      res <- collapse2plm_lag_diff(collapse::flag(x = x, n = k, shift = "time"), k = k)
+    }
+  } else {
+    ## row-wise shifting
+    if (!isTRUE(getOption("plm.fast"))) {
+      res <- lagr.pseries(x = x, k = k, ...) # base R
+    } else {
+      if (!isTRUE(getOption("plm.fast.pkg.collapse"))) stop(txt.no.collapse, call. = FALSE)
+      res <- collapse2plm_lag_diff(collapse::flag(x = x, n = k, shift = "row"), k = k)
+    }
+  }
+  res
 }
 
 #' @rdname lag.plm
 #' @export
 lead.pseries <- function(x, k = 1L, shift = c("time", "row"), ...) {
   shift <- match.arg(shift)
-  res <- if(shift == "time") leadt.pseries(x = x, k = k, ...) else leadr.pseries(x = x, k = k, ...)
-  return(res)
+  
+  if (shift == "time") {
+    if (!isTRUE(getOption("plm.fast"))) {
+      res <- leadt.pseries(x = x, k = k, ...) # base R
+    } else {
+      if (!isTRUE(getOption("plm.fast.pkg.collapse"))) stop(txt.no.collapse, call. = FALSE)
+      res <- collapse2plm_lag_diff(collapse::flag(x = x, n = -k, shift = "time"), k = k)
+    }
+  } else {
+    ## row-wise shifting
+    if (!isTRUE(getOption("plm.fast"))) {
+      res <- leadr.pseries(x = x, k = k, ...) # base R
+    } else {
+      if (!isTRUE(getOption("plm.fast.pkg.collapse"))) stop(txt.no.collapse, call. = FALSE)
+      res <- collapse2plm_lag_diff(collapse::flag(x = x, n = -k, shift = "row"), k = k)
+    }
+  }
+  res
 }
 
 #' @rdname lag.plm
 #' @exportS3Method
 diff.pseries <- function(x, lag = 1L, shift = c("time", "row"), ...) {
   shift <- match.arg(shift)
-  res <- if(shift == "time") difft.pseries(x = x, lag = lag, ...) else diffr.pseries(x = x, lag = lag, ...)
-  return(res)
+  
+  if (shift == "time") {
+    if (!isTRUE(getOption("plm.fast"))) {
+      res <- difft.pseries(x = x, lag = lag, ...) # base R
+    } else {
+      if (!isTRUE(getOption("plm.fast.pkg.collapse"))) stop(txt.no.collapse, call. = FALSE)
+      res <- collapse2plm_lag_diff(collapse::fdiff(x = x, n = lag, shift = "time"), k = lag)
+    }
+  } else {
+    ## row-wise shifting
+    if (!isTRUE(getOption("plm.fast"))) {
+      res <- diffr.pseries(x = x, lag = lag, ...) # base R
+    } else {
+      if (!isTRUE(getOption("plm.fast.pkg.collapse"))) stop(txt.no.collapse, call. = FALSE)
+      res <- collapse2plm_lag_diff(collapse::fdiff(x = x, n = lag, shift = "row"), k = lag)
+    }
+  }
+  res
 }
 
 ## lagt.pseries lagging taking the time variable into account
@@ -798,7 +846,7 @@ lagt.pseries <- function(x, k = 1L, ...) {
   else {
     rval <- alagt(x, k)
   }
-  return(rval)
+  rval
 }
 
 ## leadt.pseries(x, k) is a wrapper for lagt.pseries(x, -k)
@@ -824,16 +872,16 @@ difft.pseries <- function(x, lag = 1L, ...){
   
   lagtx <- lagt.pseries(x, k = lag) # use "time-based" lagging for difft
   
-  if(is.matrix(lagtx)) {
+  res <- if (is.matrix(lagtx)) {
     # if 'lagtx' is matrix (case length(lag) > 1):
-    # perform subtraction without pseries feature of 'x', because otherwise 
+    # perform subtraction without pseries feature of 'x', because otherwise
     # the result would be c("pseries", "matrix") which is not supported
-    res <- as.numeric(x) - lagtx
+    as.numeric(x) - lagtx
   } else {
-    res <- x - lagtx
+    x - lagtx
   }
   
-  return(res)
+  res
 }
 
 ## alagt: non-exported helper function for lagt (actual work horse),
@@ -867,17 +915,16 @@ alagt <- function(x, ak) {
     if(! anyNA(numlevtime)) time <- as.numeric(levels(time))[as.integer(time)]
     else time <- as.numeric(time)
     
-    list_id_timevar <- split(time, id, drop = TRUE)
+    list_id_timevar <- collapse::gsplit(time, id, drop = TRUE, use.names = FALSE)
     
     index_lag_ak_all_list <- sapply(list_id_timevar,
                                     function(x) match(x - ak, x, incomparables = NA),
                                     simplify = FALSE, USE.NAMES = FALSE)
     
     # translate block-wise positions to positions in full vector
-    index_lag_ak_all <- unlist(index_lag_ak_all_list, use.names = FALSE)
+    substitute_blockwise <- unlist(index_lag_ak_all_list, use.names = FALSE)
     
-    NApos <- is.na(index_lag_ak_all) # save NA positions for later
-    substitute_blockwise <- index_lag_ak_all
+    NApos <- is.na(substitute_blockwise) # save NA positions for later
     
     block_lengths <- lengths(index_lag_ak_all_list, use.names = FALSE)
     
@@ -896,7 +943,7 @@ alagt <- function(x, ak) {
     x[NApos] <- NA  # set NAs where necessary
     attributes(x) <- orig_attr # restore original names and 'pseries' class (lost by subsetting x)
   }
-  return(x)
+  x
 } # END alagt
 
 
@@ -969,7 +1016,7 @@ lagr.pseries <- function(x, k = 1L, ...) {
 leadr.pseries <- function(x, k = 1L, ...) {
     ret <- lagr.pseries(x, k = -k)
     if(length(k) > 1L) colnames(ret) <- k
-    return(ret)
+    ret
 }
 
 ## diffr: lagging row-wise
@@ -987,32 +1034,57 @@ diffr.pseries <- function(x, lag = 1L, ...) {
 
     lagrx <- lagr.pseries(x, k = lag)
     
-    if(is.matrix(lagrx)) {
+    res <- if (is.matrix(lagrx)) {
       # if 'lagrx' is matrix (case length(lag) > 1):
-      # perform subtraction without pseries feature of 'x', because otherwise 
+      # perform subtraction without pseries feature of 'x', because otherwise
       # the result would be c("pseries", "matrix") which is not supported
-      res <- as.numeric(x) - lagrx
+      as.numeric(x) - lagrx
     } else {
-      res <- x - lagrx
+      x - lagrx
     }
-    return(res)
+    
+    res
 }
 
+
 ## pdiff is (only) used in model.matrix to calculate the
-## model.matrix for FD models, works for effect = "individual" only,
+## model.matrix for FD models
+## wrapper
+pdiff <- function(x, effect = c("individual", "time"),
+                  has.intercept = FALSE,
+                  shift = c("time", "row")) {
+  
+    shift <- match.arg(shift)
+    res <- if (shift == "time")
+              pdifft(x = x,
+                     effect = effect,
+                     has.intercept = has.intercept)
+            else
+              pdiffr(x = x,
+                     effect = effect,
+                     has.intercept = has.intercept)
+    res
+}
+  
+## pdiffr works for effect = "individual" only,
 ## see model.matrix on how to call pdiff. Result is in order (id,
 ## time) for both effects
 ##
-## Performs row-wise shifting
-pdiff <- function(x, effect = c("individual", "time"), has.intercept = FALSE){
+## Performs row-wise shifting (note the 'r' in pdiffr)
+pdiffr <- function(x, effect = c("individual", "time"), has.intercept = FALSE){
   # NB: x is assumed to have an index attribute, e.g., a pseries
   #     can check with has.index(x)
+  # TODO: pdiff's usage in model.matrix is not very elegant as pdiff does its own
+  #     removal of constant columns and intercept handling which could be handled
+  #     via model.matrix.
+  
     effect <- match.arg(effect)
     cond <- as.numeric(unclass(attr(x, "index"))[[1L]]) # unclass for speed
     n <- if(is.matrix(x)) nrow(x) else length(x)
     cond <- c(NA, cond[2:n] - cond[seq_len(n-1)]) # this assumes a certain ordering
     cond[cond != 0] <- NA
     if(! is.matrix(x)){
+
         result <- c(NA , x[2:n] - x[seq_len(n-1)])
         result[is.na(cond)] <- NA
         result <- na.omit(result)
@@ -1021,13 +1093,56 @@ pdiff <- function(x, effect = c("individual", "time"), has.intercept = FALSE){
         result <- rbind(NA, x[2:n, , drop = FALSE] - x[seq_len(n-1), , drop = FALSE])
         result[is.na(cond), ] <- NA
         result <- na.omit(result)
-        result <- result[ , apply(result, 2, var) > 1E-12, drop = FALSE]
+        # remove constant columns
+        cst.col <- apply(result, 2, is.constant)
+        result <- result[ , !cst.col, drop = FALSE]
         if(has.intercept){
             result <- cbind(1, result)
             colnames(result)[1L] <- "(Intercept)"
         }
     }
+
     attr(result, "na.action") <- NULL
     result
 }
 
+## performs time-wise shifting (note the 't' in pdifft)
+pdifft <- function(x, effect = c("individual", "time"), has.intercept = FALSE) {
+  effect <- match.arg(effect)
+  x.index <- attr(x, "index")
+  x.pdf <- as.data.frame(x, make.names = FALSE)
+  x.pdf <- cbind(x.index, x.pdf)
+  x.pdf <- pdata.frame(x.pdf, drop.index = TRUE)
+  
+  if(!is.matrix(x)) {
+    # pseries case (LHS)
+    res <- diff(x, effect = effect, shift = "time")
+    res <- subset_pseries(res, !is.na(res)) # TODO: use [.pseries (pseries subsetting) once implemented
+  } else {
+    # matrix case (RHS)
+    res <- apply(x, 2, function(col) diff(add_pseries_features(col, x.index),
+                                          effect = effect, shift = "time"))
+    res <- na.omit(res)
+    # if intercept is requested, set intercept column to 1 as it was diff'ed out
+    if(has.intercept) res[ , 1L] <- 1L
+  }
+  res
+}
+
+### non-exported helper function to align collapse::flag to plm's original lag
+## https://github.com/SebKrantz/collapse/issues/183 and there is no factor matrix
+## https://stackoverflow.com/a/28724756/4640346
+collapse2plm_lag_diff <- function(x, k) {
+  if(is.matrix(x)) {
+    dim.nam1 <- dimnames(x)[[1L]]
+    if(inherits(x, "factor")) {
+      # cater for collapse's special factor matrix case
+      x <- matrix(as.character(x), ncol = dim(x)[[2L]])
+      dimnames(x)[[1L]] <- dim.nam1
+    }
+    class(x) <- NULL
+    attr(x, "index") <- NULL
+    dimnames(x)[[2L]] <- k
+  }
+  x
+}